requirements.txt
langchain[openai]
foundry-local-sdk
botbuilder-integration-aiohttp